Cleaned up deprecated fetching / querying functions and state that are no longer needed.
