+++ backend/api/routes.go
@@ -9,7 +9,6 @@ import (
 
 func InitRoutes(e *echo.Echo) {
 	e.GET("/otel/span-service-names", otel.GetDistinctServiceNames)
-	e.GET("/workflow-spans-e2e/:serviceName", otel.GetWorkflowE2E)
 	e.GET("/otel/service/:serviceName/root-spans", otel.GetRootSpans)
 	e.GET("/otel/service/:serviceName/root-spans-filtered", otel.GetRootSpansFiltered)
 	e.GET("/otel/trace/:traceId/nested-spans", otel.GetNestedSpans)
