+++ frontend/src/features/junjo-data/workflow-detail/SpanExceptionsList.tsx
@@ -1,6 +1,6 @@
 import { useRef } from 'react'
 import { useAppDispatch } from '../../../root-store/hooks'
-import { OtelSpan } from '../../otel/schemas/schemas'
+import { OtelSpan } from '../../traces/schemas/schemas'
 import { SpanIconConstructor } from '../span-lists/determine-span-icon'
 import { WorkflowDetailStateActions } from './store/slice'
