+++ frontend/src/features/junjo-data/workflow-detail/store/selectors.ts
@@ -1,0 +1,5 @@
+import { RootState } from '../../../../root-store/store'
+
+// Selectors - Workflow Detail
+export const selectActiveSetStateEvent = (state: RootState) => state.workflowDetailState.activeSetStateEvent
+export const selectWorkflowDetailActiveSpan = (state: RootState) => state.workflowDetailState.activeSpan
