+++ frontend/src/features/prompt-playground/PromptPlaygroundPage.tsx
@@ -1,6 +1,6 @@
 import { Link, useParams } from 'react-router'
 import { useEffect, useState } from 'react'
-import { OtelSpan } from '../otel/schemas/schemas'
+import { OtelSpan } from '../traces/schemas/schemas'
 import { API_HOST } from '../../config'
 import { useAppDispatch, useAppSelector } from '../../root-store/hooks'
 import { geminiTextRequest } from './fetch/gemini-text-request'
