+++ frontend/src/features/traces/TraceDetails.tsx
@@ -1,6 +1,6 @@
 import { Link, useParams, useNavigate } from 'react-router'
 import { useEffect, useState } from 'react'
-import { OtelSpan } from '../otel/schemas/schemas'
+import { OtelSpan } from '../traces/schemas/schemas'
 import { API_HOST } from '../../config'
 import NestedOtelSpans from './NestedOtelSpans'
 import SpanAttributesPanel from './SpanAttributesPanel'
