+++ frontend/src/features/traces/schemas/attribute-schemas-openinference.ts
