+++ frontend/src/features/traces/store/listeners.ts
@@ -2,9 +2,29 @@ import { createListenerMiddleware } from '@reduxjs/toolkit/react'
 import { AppDispatch, RootState } from '../../../root-store/store'
 import { TracesStateActions } from './slice'
 import { getTraceSpans } from '../fetch/get-trace-spans'
+import { fetchServiceNames } from '../fetch/get-service-names'
 
 export const otelStateListenerMiddleware = createListenerMiddleware()
 const startListener = otelStateListenerMiddleware.startListening.withTypes<RootState, AppDispatch>()
+
+startListener({
+  actionCreator: TracesStateActions.fetchServiceNames,
+  effect: async (_action, listenerApi) => {
+    // Clear errors and set loading
+    listenerApi.dispatch(TracesStateActions.setServiceNamesError(false))
+    listenerApi.dispatch(TracesStateActions.setServiceNamesLoading(true))
+
+    // Fetch the data
+    try {
+      const data = await fetchServiceNames()
+      listenerApi.dispatch(TracesStateActions.setServiceNamesData(data))
+    } catch (error) {
+      listenerApi.dispatch(TracesStateActions.setServiceNamesError(true))
+    } finally {
+      listenerApi.dispatch(TracesStateActions.setServiceNamesLoading(false))
+    }
+  },
+})
 
 startListener({
   actionCreator: TracesStateActions.fetchSpansByTraceId,
@@ -12,7 +32,7 @@ startListener({
     const { traceId } = action.payload
     if (!traceId) throw new Error('No traceId provided')
 
-    const { loading } = getState().otelState.workflows
+    const loading = getState().tracesState.loading
     if (loading) return
 
     // // Cache busting logic
