+++ frontend/src/features/traces/store/slice.ts
@@ -1,8 +1,13 @@
 import { createSlice } from '@reduxjs/toolkit'
 import type { PayloadAction } from '@reduxjs/toolkit'
-import { OtelSpan } from '../../otel/schemas/schemas'
+import { OtelSpan } from '../../traces/schemas/schemas'
 
 interface TracesState {
+  serviceNames: {
+    data: string[]
+    loading: boolean
+    error: boolean
+  }
   traceSpans: {
     [traceId: string]: OtelSpan[]
   }
@@ -11,6 +16,11 @@ interface TracesState {
 }
 
 const initialState: TracesState = {
+  serviceNames: {
+    data: [],
+    loading: false,
+    error: false,
+  },
   traceSpans: {},
   loading: false,
   error: false,
@@ -24,7 +34,22 @@ export const tracesSlice = createSlice({
     fetchSpansByTraceId: (_state, _action: PayloadAction<{ traceId: string | undefined }>) => {
       // Handled by listener middleware
     },
+    fetchServiceNames: (_state) => {
+      // Handled by listener middleware
+    },
 
+    // Service Names Actions
+    setServiceNamesData: (state, action: PayloadAction<string[]>) => {
+      state.serviceNames.data = action.payload
+    },
+    setServiceNamesLoading: (state, action: PayloadAction<boolean>) => {
+      state.serviceNames.loading = action.payload
+    },
+    setServiceNamesError: (state, action: PayloadAction<boolean>) => {
+      state.serviceNames.error = action.payload
+    },
+
+    // Traces Data Actions
     setTracesData: (state, action: PayloadAction<{ traceId: string; data: OtelSpan[] }>) => {
       state.traceSpans[action.payload.traceId] = action.payload.data
     },
