+++ frontend/src/features/traces/utils/span-utils.ts
